chore: place overrides in declaration.d.ts file for accurate ts autocomplete/definitions
